Threadz-V1/backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic[email]==2.12.5
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.3.0
pillow==12.1.1
sentry-sdk[fastapi]==2.19.2
python-dotenv==1.0.1
razorpay==1.4.2
redis==5.2.0
celery==5.4.0
httpx==0.28.1
boto3==1.34.144
stability-sdk==0.8.0
openai==1.52.0
fastapi-mail==1.4.1
jinja2==3.1.4
slowapi==0.1.9
pytest==8.3.2
pytest-cov==6.0.0
pytest-asyncio==0.24.0
aiosmtplib==3.0.2
clamd==1.0.2
